Fix user vimrc file check
 - check file's existence prior to the backup attempt
 - enable syntax
 - prefer curl over wget; wget may not be available in minimal distros

+++ setup-vim-familiar.sh
@@ -8,9 +8,18 @@ mkdir -p ~/.vim/autoload ~/.vim/bundle || ( echo "Could not create dirs\n" && \
 	exit 1 )
 curl -LSso ~/.vim/autoload/pathogen.vim https://tpo.pe/pathogen.vim 
 
-cp -f ~/.vimrc ~/.vimrc.BAK
-echo "\" Update runtimepaths" >> ~/.vimrc
-echo "execute pathogen#infect()" >> ~/.vimrc
+# Backup user vimrc if it exists
+if [ -s ~/.vimrc ]
+then
+	cp -f ~/.vimrc ~/.vimrc.BAK
+fi
+
+# Add pathogen call to vimrc
+if ! grep -qe "execute pathogen#infect()" ~/.vimrc &> /dev/null
+then
+	echo "\" Update runtimepaths" >> ~/.vimrc
+	echo "execute pathogen#infect()" >> ~/.vimrc
+fi
 
 # Place the familiar plugin 
 mkdir -p ~/.vim/bundle/familiar/plugin || ( echo "Could not create dirs\n" && \
@@ -18,9 +27,9 @@ mkdir -p ~/.vim/bundle/familiar/plugin || ( echo "Could not create dirs\n" && \
 cp -f ./plugin/familiar.vim ~/.vim/bundle/familiar/plugin/
 
 # Pull and setup TPope's vim-surround
-mkdir -p ~/.vim/bundle/vim-surround/plugin || ( echo "Could not create dirs\n" && \
-	exit 1 )
-wget -q -O ~/.vim/bundle/vim-surround/plugin/surround.vim \
+mkdir -p ~/.vim/bundle/vim-surround/plugin || ( echo "Could not create dirs\n"\
+	&& exit 1 )
+curl -LSso ~/.vim/bundle/vim-surround/plugin/surround.vim \
 https://raw.githubusercontent.com/tpope/vim-surround/master/plugin/surround.vim
 
 exit 0
